requirements.txt
nose==1.3.7
python-frontmatter==0.4.5
PyYAML==5.1.2
Pillow==6.2.1
